requirements.txt
hidapi
tkinter